[Standard] Changed: Move all transform test logic into the shared prelude
