+++ projects/dxt-lossless-transform-bc1/src/transforms/standard/transform/avx2.rs
@@ -415,9 +415,6 @@ pub unsafe fn shuffle_permute_unroll_2_with_separate_pointers(
 mod tests {
     use super::*;
     use crate::test_prelude::*;
-    use crate::transforms::standard::untransform::untransform;
-
-    type PermuteFn = unsafe fn(*const u8, *mut u8, usize);
 
     #[rstest]
     #[case(shuffle_permute, "shuffle_permute")]
@@ -425,33 +422,11 @@ mod tests {
     #[case(permute, "permute")]
     #[case(permute_unroll_2, "permute unroll 2")]
     #[case(gather, "gather")]
-    fn avx2_transform_roundtrip(#[case] permute_fn: PermuteFn, #[case] impl_name: &str) {
-        if !dxt_lossless_transform_common::cpu_detect::has_avx2() {
+    fn avx2_transform_roundtrip(#[case] permute_fn: TransformFn, #[case] impl_name: &str) {
+        if !has_avx2() {
             return;
         }
 
-        for num_blocks in 1..=128 {
-            let input = generate_bc1_test_data(num_blocks);
-            let len = input.len();
-            let mut transformed = vec![0u8; len];
-            let mut reconstructed = vec![0u8; len];
-            unsafe {
-                permute_fn(
-                    input.as_ptr(),
-                    transformed.as_mut_ptr(),
-                    len,
-                );
-                untransform(
-                    transformed.as_ptr(),
-                    reconstructed.as_mut_ptr(),
-                    len,
-                );
-            }
-            assert_eq!(
-                reconstructed.as_slice(),
-                input.as_slice(),
-                "Mismatch AVX2 roundtrip {impl_name}",
-            );
-        }
+        run_standard_transform_roundtrip_test(permute_fn, 128, impl_name);
     }
 }
