+++ projects/dxt-lossless-transform-bc1/src/transforms/standard/transform/sse2.rs
@@ -312,36 +312,17 @@ pub unsafe fn shufps_unroll_4_with_separate_pointers(
 mod tests {
     use super::*;
     use crate::test_prelude::*;
-    use crate::transforms::standard::untransform::untransform;
-
-    type PermuteFn = unsafe fn(*const u8, *mut u8, usize);
 
     #[rstest]
     #[case(punpckhqdq_unroll_4, "SSE2 punpckhqdq unroll-4")]
     #[case(punpckhqdq_unroll_2, "SSE2 punpckhqdq unroll-2")]
     #[case(shufps_unroll_2, "SSE2 shuffle unroll-2")]
     #[case(shufps_unroll_4, "SSE2 shuffle unroll-4")]
-    fn sse2_transform_roundtrip(#[case] permute_fn: PermuteFn, #[case] impl_name: &str) {
-        if !dxt_lossless_transform_common::cpu_detect::has_sse2() {
+    fn sse2_transform_roundtrip(#[case] permute_fn: TransformFn, #[case] impl_name: &str) {
+        if !has_sse2() {
             return;
         }
 
-        for num_blocks in 1..=512 {
-            let input = generate_bc1_test_data(num_blocks);
-            let len = input.len();
-            let mut transformed = vec![0u8; len];
-            let mut reconstructed = vec![0u8; len];
-
-            unsafe {
-                permute_fn(input.as_ptr(), transformed.as_mut_ptr(), len);
-                untransform(transformed.as_ptr(), reconstructed.as_mut_ptr(), len);
-            }
-
-            assert_eq!(
-                reconstructed.as_slice(),
-                input.as_slice(),
-                "Mismatch {impl_name} roundtrip for {num_blocks} blocks",
-            );
-        }
+        run_standard_transform_roundtrip_test(permute_fn, 512, impl_name);
     }
 }
